+++ server/ConfigDI/Interfaces.h
@@ -20,7 +20,7 @@ class IGlobalGridCfg
 public:
   virtual ~IGlobalGridCfg() = default;
 
-  virtual uint8_t  gridSize()    const = 0;
+  virtual uint16_t gridSize()    const = 0;
   virtual uint16_t cellWidthKm() const = 0;
 };
